Corrected Master table

+++ src/main/java/com/hcl/fsc/repositories/CandidateRepository.java
@@ -1,12 +1,14 @@
 package com.hcl.fsc.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Repository;
+
 import com.hcl.fsc.entities.Candidate;
 
 
 
 @Repository
-public interface CandidateRepository extends JpaRepository<Candidate, Integer>{
+public interface CandidateRepository extends JpaRepository<Candidate, Long>{
 	
 }
